pyproject.toml
[tool.poetry]
name = "bicycle"
version = "0.1.0"
description = "A bicycle for the mind."
authors = ["Devon Peticolas <devon.peticolas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.25.0"
Flask = "^2.2.2"
python-dotenv = "^0.21.0"
pandas = "^1.5.1"
sqlparse = "^0.4.3"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
mypy = "^0.982"

[tool.mypy]
python_version = 3.10
ignore_missing_imports = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
no_warn_no_return = true
warn_unreachable = true
allow_redefinition = false
strict_equality = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
